requirements.txt
streamlit
openai
PyPDF2
docx2txt
python-docx